Backend/requirements.txt
aniso8601==9.0.1
attrs==22.2.0
click==8.1.3
configparser==5.3.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
flask-restx==1.0.6
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
mysql-client==0.0.1
mysql-connector-python==8.0.32
mysqlclient==2.1.1
protobuf==3.20.3
PyMySQL==1.0.2
pyrsistent==0.19.3
python-decouple==3.7
pytz==2022.7.1
six==1.16.0
SQLAlchemy==2.0.3
typing_extensions==4.4.0
Werkzeug==2.2.2
